+++ flowserv/config.py
@@ -184,13 +184,21 @@ FLOWSERV_WEBAPP = 'FLOWSERV_WEBAPP'
 """Environment variables that are used to configure the file store that is used
 to maintain files for workflow templates, user group uploads, and workflow runs.
 """
-# Name of the class that implements the file store interface
-FLOWSERV_FILESTORE_CLASS = 'FLOWSERV_FILESTORE_CLASS'
-# Name of the module that contains the file store implementation
-FLOWSERV_FILESTORE_MODULE = 'FLOWSERV_FILESTORE_MODULE'
+# Identifier of the file store class.
+FLOWSERV_FILESTORE = 'FLOWSERV_FILESTORE'
+# Identifier of the bucket type if the BucketStore is used.
+FLOWSERV_FILESTORE_BUCKETTYPE = 'FLOWSERV_FILESTORE_BUCKETTYPE'
+# Accepted values for the file store class.
+FILESTORE_BUCKET = 'bucket'
+FILESTORE_FS = 'fs'
+# Accepted values for bucket type.
+BUCKET_FS = 'fs'  # Disk bucket for test purposes
+BUCKET_GC = 'gc'  # Google Cloud File Store
+BUCKET_MEM = 'mem'  # In-memory bucket for test purposes.
+BUCKET_S3 = 's3'  # AWS S3
 
 """Environment variable for unique bucket identifier."""
-FLOWSERV_S3BUCKET = 'FLOWSERV_S3BUCKET'
+FLOWSERV_BUCKET = 'FLOWSERV_BUCKET'
 
 
 # --
@@ -312,9 +320,9 @@ class Config(dict):
         -------
         flowserv.config.Config
         """
-        self[FLOWSERV_FILESTORE_CLASS] = 'BucketStore'
-        self[FLOWSERV_FILESTORE_MODULE] = 'flowserv.model.files.s3'
-        self[FLOWSERV_S3BUCKET] = bucket
+        self[FLOWSERV_FILESTORE] = FILESTORE_BUCKET
+        self[FLOWSERV_FILESTORE_BUCKETTYPE] = BUCKET_S3
+        self[FLOWSERV_BUCKET] = bucket
         return self
 
     def token_timeout(self, timeout: int) -> Config:
@@ -448,9 +456,9 @@ ENV = [
     (FLOWSERV_CLIENT, LOCAL_CLIENT, None),
     (FLOWSERV_DB, None, None),
     (FLOWSERV_WEBAPP, 'False', to_bool),
-    (FLOWSERV_FILESTORE_CLASS, None, None),
-    (FLOWSERV_FILESTORE_MODULE, None, None),
-    (FLOWSERV_S3BUCKET, None, None)
+    (FLOWSERV_FILESTORE, None, None),
+    (FLOWSERV_FILESTORE_BUCKETTYPE, None, None),
+    (FLOWSERV_BUCKET, None, None)
 ]